nv-repos/dataflow_apache_beam/requirements.txt
google-api-core==1.16.0
google-apitools==0.5.28
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.24.0
google-cloud-bigtable==1.0.0
google-cloud-core==1.1.0
google-cloud-datastore==1.7.4
google-cloud-dlp==0.13.0
google-cloud-language==1.3.0
google-cloud-pubsub==1.0.2
google-cloud-spanner==1.13.0
google-cloud-videointelligence==1.13.0
google-cloud-vision==0.42.0
google-pasta==0.2.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0

